Moved MediaFinder formwidget from the CMS module to the System module
