+++ modules/backend/formwidgets/mediafinder/assets/js/mediafinder.js
