+++ modules/backend/formwidgets/mediafinder/partials/_image_single.htm
