+++ modules/cms/formwidgets/MediaFinder.php
@@ -1,10 +1,6 @@
 <?php namespace Cms\FormWidgets;
 
-use Lang;
-use ApplicationException;
-use Cms\Classes\MediaLibrary;
-use Backend\Classes\FormField;
-use Backend\Classes\FormWidgetBase;
+use Backend\FormWidgets\MediaFinder as BackendMediaFinder;
 
 /**
  * Media Finder
@@ -17,102 +13,20 @@ use Backend\Classes\FormWidgetBase;
  *
  * @package october\cms
  * @author Alexey Bobkov, Samuel Georges
+ * @deprecated Use Backend\FormWidgets\MediaFinder. Remove if year >= 2020.
  */
-class MediaFinder extends FormWidgetBase
+class MediaFinder extends BackendMediaFinder
 {
-    //
-    // Configurable properties
-    //
-
     /**
-     * @var string Prompt to display if no record is selected.
+     * Constructor.
      */
-    public $prompt = 'cms::lang.mediafinder.default_prompt';
-
-    /**
-     * @var string Display mode for the selection. Values: file, image.
-     */
-    public $mode = 'file';
-
-    /**
-     * @var int Preview image width
-     */
-    public $imageWidth = null;
-
-    /**
-     * @var int Preview image height
-     */
-    public $imageHeight = null;
-
-    //
-    // Object properties
-    //
-
-    /**
-     * @inheritDoc
-     */
-    protected $defaultAlias = 'media';
-
-    /**
-     * @inheritDoc
-     */
-    public function init()
+    public function __construct()
     {
-        $this->fillFromConfig([
-            'mode',
-            'prompt',
-            'imageWidth',
-            'imageHeight'
-        ]);
+        traceLog('FormWidget Cms\FormWidgets\MediaFinder has been deprecated, use Backend\FormWidgets\MediaFinder instead.');
 
-        if ($this->formField->disabled) {
-            $this->previewMode = true;
-        }
-    }
+        $this->assetPath = '/modules/backend/formwidgets/mediafinder/assets';
+        $this->viewPath = base_path('/modules/backend/formwidgets/mediafinder/partials')
 
-    /**
-     * @inheritDoc
-     */
-    public function render()
-    {
-        $this->prepareVars();
-
-        return $this->makePartial('mediafinder');
-    }
-
-    /**
-     * Prepares the list data
-     */
-    public function prepareVars()
-    {
-        $value = $this->getLoadValue();
-        $this->vars['value'] = $value;
-        $this->vars['imageUrl'] = $value ? MediaLibrary::url($value) : '';
-        $this->vars['field'] = $this->formField;
-        $this->vars['prompt'] = str_replace('%s', '<i class="icon-folder"></i>', trans($this->prompt));
-        $this->vars['mode'] = $this->mode;
-        $this->vars['imageWidth'] = $this->imageWidth;
-        $this->vars['imageHeight'] = $this->imageHeight;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getSaveValue($value)
-    {
-        if ($this->formField->disabled || $this->formField->hidden) {
-            return FormField::NO_SAVE_DATA;
-        }
-
-        return $value;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    protected function loadAssets()
-    {
-        $this->addJs('js/mediafinder.js', 'core');
-        $this->addCss('css/mediafinder.css', 'core');
+        parent::__construct(...func_get_args());
     }
 }
